requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy
opencv-python
pillow
yacs
tqdm
tensorboardX
matplotlib
